GamesCategory -> GamesList
