+++ client/src/pages/Games.tsx
@@ -1,18 +1,12 @@
-import React, {useEffect} from 'react';
-import GamesCategory from "../components/GamesCategory";
+import React from 'react';
+import GamesList from "../components/GamesList";
 import MainLayout from "../components/layouts/MainLayout";
-import {useAppDispatch, useAppSelector} from "../hooks/redux";
 
 const Games = () => {
-    const {games, isLoading, error} = useAppSelector(state => state.game)
-    const dispatch = useAppDispatch()
-
     return (
         <MainLayout>
             <div className={"w-full flex flex-col"}>
-                <GamesCategory title={"Популярные"} type={"popular"} games={games} isLoading={isLoading} error={error}/>
-                <GamesCategory title={"Распродажа"} type={"sale"} games={games} isLoading={isLoading} error={error}/>
-                <GamesCategory title={"Новинки"} type={"new"} games={games} isLoading={isLoading} error={error}/>
+                <GamesList/>
             </div>
         </MainLayout>
     );
